+++ src/data/daomos.js
@@ -1,39 +1,331 @@
+import daomos1 from 'assets/images/projects/daomos/daomos1.png';
+import daomos2 from 'assets/images/projects/daomos/daomos1.png';
+import daomos3 from 'assets/images/projects/daomos/daomos1.png';
+import daomos4 from 'assets/images/projects/daomos/daomos4.png';
+import daomos5 from 'assets/images/projects/daomos/daomos5.png';
+import daomos6 from 'assets/images/projects/daomos/daomos6.png';
+import daomos7 from 'assets/images/projects/daomos/daomos7.png';
+import daomos8 from 'assets/images/projects/daomos/daomos1.png';
+
 const daomos = {
     id: 'daomos',
     behanceLink: 'https://behance.net',
     title: 'Daomos',
     en: {
-        subtitle: 'A working space web3 based',
-        overview: [
-            'Project created from scratch, in 4 days, in collaboration with Tadeau Carabias.',
-            'The challenge was to create a responsive online platform for a magazine to meet the needs and goals of users. The goal was to deliver a consistent, branded, and engaging experience and our main focus were on UI Design.'
-        ],
+        subtitle: 'Dapp for DAOs',
+        overview:
+            'For this project, I worked with Jaoa Pereira (Build up Labs), as the stakeholder and the research for this project was teamwork with Catarina Silva. The main challenge was to create a tool to enhance DAOs and keep contributing to the automatization and decentralization of the web3 ecosystem.',
         scope: 'User Research, analyzing, user flow, information architecture, wireframing, mock-up, design system, prototyping, and usability testing.',
-        categories: 'Website/Mobile ',
-        brief: 'Daomos is an emergent art magazine to promote new and upcoming artists and review the work of established ones, with the intent to democratize art and enable easier access for everyone both artist and viewers',
-        research: [
-            'Starting our creative process by choosing an already created user persona (Paula) allowed us to think about this editorial project, which mutated and evolved from the beginning excitingly and organically. By having specific characteristics about Paula, we could orient our research.',
-            'At first, we searched the web for different digital and physical magazines to understand how this type of market works and what people want to consume when they decide to read an art magazine.',
-            'After visiting and comparing different sites, the information they offer, how content is organized, their aesthetics, their accessibility, and even the frequency of content, we started thinking of questions for the interviews that would allow us to achieve our goal.'
-        ],
-        visualConcept:
-            'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.',
-        prototype:
-            'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.'
+        categories: 'Mobile',
+        post: [
+            {
+                sectionNumber: '01',
+                id: 'brief',
+                content: [
+                    {
+                        type: 'paragraph',
+                        content:
+                            'Daomos is a  decentralized platform to help day-to-day DAOs.It facilitates organization and communication between participants who want to contribute to a decentralized world and revolutionize the future of work. An app that generates spaces where people build communities gets to know people who share the same interests and values and connects people with cool and meaningful web3-based projects around the world.',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'image',
+                        content: { src: daomos1, alt: 'Daomos 1', size: 'auto' }
+                    }
+                ]
+            },
+            {
+                sectionNumber: '02',
+                id: 'research',
+                content: [
+                    {
+                        type: 'paragraph',
+                        content:
+                            'To create a digital product that lives on the web3, we first had to conduct extensive desktop research on this new world born in 2018.',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'We need to understand what it means for a product to live on the blockchain and the implications and benefits of doing it in this ecosystem and not in the current one ( web2).',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'Besides looking for articles, we participated in different online talks on the subject, approached face-to-face web3 groups in Lisbon, and interviewed an expert in web3 design to start understanding the users of web3.',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'We also deep dig into the Decentralized Autonomous Organization concept to understand what they are, how they work, how it looks like the actual DAO tooling landscape, and how we can boost it with a unique product.',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'video',
+                        content: { src: 'https://www.youtube.com/embed/AKBmmVCL87U', alt: 'Daomos 2' }
+                    },
+                    {
+                        type: 'paragraph',
+                        content: 'We also analyzed our web3 competitors.',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'video',
+                        content: { src: 'https://www.youtube.com/embed/nvhYqMqH-bQ', alt: 'Daomos 3' }
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'Finally, we conducted 36 surveys to understand why or why not a user would choose a web3-based product. Here are some thoughts of our users.',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'image',
+                        content: { src: daomos4, alt: 'Daomos 4', size: 'auto' }
+                    },
+                    {
+                        type: 'paragraph',
+                        content: 'Some insights of our research',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'image',
+                        content: { src: daomos5, alt: 'Daomos 5', size: 'auto' }
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'We have learned that in products like this it is crucial to make the greatest effort to develop a great user experience and thus be able to invite more people to join something still unknown and often overwhelming.',
+                        size: '200',
+                        weight: '600',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'We need a DAPP where users own and control their own data and interact directly with one another, free from surveillance or censorship, while still having access to the same breadth and quality of services.',
+                        size: '200',
+                        weight: '600',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'How might we create a scheduling tool abstracted away as much technical excess possible and be much more user-friendly?',
+                        size: '200',
+                        weight: '600',
+                        style: 'normal'
+                    }
+                ]
+            },
+            {
+                sectionNumber: '03',
+                id: 'visualConcept',
+                content: [
+                    {
+                        type: 'paragraph',
+                        content: 'The UI',
+                        size: '200',
+                        weight: '600',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'images-slide',
+                        size: 'auto',
+                        content: [
+                            { src: daomos6, alt: 'Daomos 6' },
+                            { src: daomos7, alt: 'Daomos 7' }
+                        ]
+                    }
+                ]
+            },
+            {
+                sectionNumber: '04',
+                id: 'prototype',
+                content: [
+                    {
+                        type: 'video',
+                        content: { src: daomos8, alt: 'Daomos 8' }
+                    }
+                ]
+            }
+        ]
     },
     es: {
-        subtitle: 'A working space web3 based',
+        subtitle: 'Una Dapp para DAOs',
         overview:
-            'Daomos is an emergent art magazine to promote new and upcoming artists and review the work of established ones, with the intent to democratize art and enable easier access for everyone both artist and viewers.',
-        focus: 'Experience and Interface Design',
-        categories: 'Website/Mobile ',
-        brief: 'Daomos is a collaborative project,  done with Tadeu Carabias in  is a  magazine oriented to a young audience. Our public intended to be an audience who enjoys knowing new and emergent artists, who like to know what happens in the artistic environment, is aware of new trends, and receives recommendations on new artistic creations to continue feeding their artistic interest spirit.',
-        research:
-            'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.',
-        visualConcept:
-            'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.',
-        prototype:
-            'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.'
+            'Para este proyecto, trabajé con Jaoa Pereira (Build up Labs), como parte interesada. La investigación para este proyecto fue un trabajo en equipo con Catarina Silva. El principal desafío fue crear una herramienta para mejorar los DAOs y seguir contribuyendo a la automatización y descentralización del ecosistema web3.',
+        scope: 'Investigación de usuarios, análisis, flujo de usuarios, arquitectura de la información, wireframing, mock-up, sistema de diseño, prototipado y pruebas de usabilidad.',
+        categories: 'Movil',
+        post: [
+            {
+                sectionNumber: '01',
+                id: 'brief',
+                content: [
+                    {
+                        type: 'paragraph',
+                        content:
+                            'Daomos es una plataforma descentralizada para ayudar el el día a día de los DAOs. Facilita la organización y la comunicación entre los participantes que quieren contribuir a un mundo descentralizado y revolucionar el futuro del trabajo. Una aplicación que genera espacios donde la gente construye comunidades, conoce a personas que comparten los mismos intereses y valores y conecta a la gente de todo el mundo con proyectos geniales y significativos basados en la web3.',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'image',
+                        content: { src: daomos1, alt: 'Daomos 1', size: 'auto' }
+                    }
+                ]
+            },
+            {
+                sectionNumber: '02',
+                id: 'research',
+                content: [
+                    {
+                        type: 'paragraph',
+                        content:
+                            'Para crear un producto digital que viva en la web3, primero tuvimos que realizar una amplia investigación de escritorio sobre este nuevo mundo nacido en 2018.',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'Necesitamos entender lo que significa que un producto viva en el blockchain y las implicaciones y beneficios de hacerlo en este ecosistema y no en el actual ( web2).',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'Además de buscar artículos, participamos en diferentes charlas online sobre el tema, nos acercamos a grupos presenciales de web3 en Lisboa, y entrevistamos a un experto en diseño de web3 para empezar a entender a los usuarios de web3.',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'También profundizamos en el concepto de Organización Autónoma Descentralizada para entender qué son, cómo funcionan, cómo es el panorama actual de las herramientas DAO, y cómo podemos potenciarlo con un producto único.',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'video',
+                        content: { src: 'https://www.youtube.com/embed/AKBmmVCL87U', alt: 'Daomos 2' }
+                    },
+                    {
+                        type: 'paragraph',
+                        content: 'También analizamos a nuestros competidores de la web3.',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'video',
+                        content: { src: 'https://www.youtube.com/embed/nvhYqMqH-bQ', alt: 'Daomos 3' }
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'Por último, realizamos 36 encuestas para entender por qué un usuario elegiría o no un producto basado en la web3. Estas son algunas de las opiniones de nuestros usuarios.',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'image',
+                        content: { src: daomos4, alt: 'Daomos 4', size: 'auto' }
+                    },
+                    {
+                        type: 'paragraph',
+                        content: 'Algunos datos de nuestra investigación',
+                        size: '200',
+                        weight: '500',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'image',
+                        content: { src: daomos5, alt: 'Daomos 5', size: 'auto' }
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'Hemos aprendido que en productos como éste es crucial hacer el mayor esfuerzo para desarrollar una gran experiencia de usuario y así poder invitar a más personas a unirse a algo todavía desconocido y a menudo abrumador.',
+                        size: '200',
+                        weight: '600',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'Necesitamos una DAPP en la que los usuarios sean dueños y controlen sus propios datos e interactúen directamente entre ellos, sin vigilancia ni censura, pero con la misma amplitud y calidad de servicios.',
+                        size: '200',
+                        weight: '600',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            '¿Cómo podríamos crear una herramienta de programación que se abstraiga de todo el exceso técnico posible y sea mucho más fácil de usar?',
+                        size: '200',
+                        weight: '600',
+                        style: 'normal'
+                    }
+                ]
+            },
+            {
+                sectionNumber: '03',
+                id: 'visualConcept',
+                content: [
+                    {
+                        type: 'paragraph',
+                        content: 'The UI',
+                        size: '200',
+                        weight: '600',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'images-slide',
+                        size: 'auto',
+                        content: [
+                            { src: daomos6, alt: 'Daomos 6' },
+                            { src: daomos7, alt: 'Daomos 7' }
+                        ]
+                    }
+                ]
+            },
+            {
+                sectionNumber: '04',
+                id: 'prototype',
+                content: [
+                    {
+                        type: 'video',
+                        content: { src: daomos8, alt: 'Daomos 8' }
+                    }
+                ]
+            }
+        ]
     }
 };