+++ src/data/rito.js
@@ -1,9 +1,9 @@
-import rito1 from 'assets/images/rito1.jpg';
-import rito2 from 'assets/images/rito2.jpg';
-import rito3 from 'assets/images/rito3.jpg';
-import rito4 from 'assets/images/rito4.jpg';
-import rito5 from 'assets/images/rito5.jpg';
-import rito6 from 'assets/images/rito1.jpg';
+import rito1 from 'assets/images/projects/rito/rito1.jpg';
+import rito2 from 'assets/images/projects/rito/rito2.jpg';
+import rito3 from 'assets/images/projects/rito/rito3.jpg';
+import rito4 from 'assets/images/projects/rito/rito4.jpg';
+import rito5 from 'assets/images/projects/rito/rito5.jpg';
+import rito6 from 'assets/images/projects/rito/rito1.jpg';
 
 const rito = {
     id: 'rito',
@@ -16,7 +16,7 @@ const rito = {
             'The challenge was to create a responsive online platform for a magazine to meet the needs and goals of users. The goal was to deliver a consistent, branded, and engaging experience and our main focus were on UI Design.'
         ],
         scope: 'User Research, analyzing, user flow, information architecture, wireframing, mock-up, design system, prototyping, and usability testing.',
-        categories: 'Website/Mobile ',
+        categories: 'Website/Mobile',
         post: [
             {
                 sectionNumber: '01',
@@ -27,12 +27,12 @@ const rito = {
                         content:
                             'Rito is an emergent art magazine to promote new and upcoming artists and review the work of established ones, with the intent to democratize art and enable easier access for everyone both artist and viewers',
                         size: '200',
-                        weight: '400',
+                        weight: '500',
                         style: 'normal'
                     },
                     {
                         type: 'image',
-                        content: { src: rito1, alt: 'Rito 1' }
+                        content: { src: rito1, alt: 'Rito 1', size: 'auto' }
                     }
                 ]
             },
@@ -45,7 +45,7 @@ const rito = {
                         content:
                             'At first, we searched the web for different digital and physical magazines to understand how this type of market works and what people want to consume when they decide to read an art magazine.',
                         size: '200',
-                        weight: '400',
+                        weight: '500',
                         style: 'normal'
                     },
                     {
@@ -53,7 +53,7 @@ const rito = {
                         content:
                             'After visiting and comparing different sites, the information they offer, how content is organized, their aesthetics, their accessibility, and even the frequency of content, we started thinking of questions for the interviews that would allow us to achieve our goal.',
                         size: '200',
-                        weight: '400',
+                        weight: '500',
                         style: 'normal'
                     },
                     {
@@ -67,7 +67,7 @@ const rito = {
                         type: 'paragraph',
                         content: 'We interviewed five people and this is what we learned from them.',
                         size: '200',
-                        weight: '400',
+                        weight: '500',
                         style: 'normal'
                     },
                     {
@@ -79,7 +79,7 @@ const rito = {
                         content:
                             'On the other hand, the surveys allow us to learn that magazines are a product that is not consumed daily, but more sporadically during the week. Also, 100% of the time, they are accessed by recommendation on social networks. Finally, people consume them because they look for inspiration and recommendations.',
                         size: '200',
-                        weight: '400',
+                        weight: '500',
                         style: 'normal'
                     },
                     {
@@ -101,7 +101,7 @@ const rito = {
                         content:
                             'We choose a  colorful palette since colors are representative of Latin America.Our name was born after brainstorming a lot. We thought about our interviews, what we heard about them, and reflected on what we understood about art.',
                         size: '200',
-                        weight: '400',
+                        weight: '500',
                         style: 'normal'
                     },
                     {
@@ -116,7 +116,7 @@ const rito = {
                 content: [
                     {
                         type: 'video',
-                        content: { src: rito6, alt: 'Rito 6' }
+                        content: { src: 'https://www.youtube.com/embed/PsXkEXV3cgA', alt: 'Rito 6' }
                     }
                 ]
             }
@@ -125,11 +125,11 @@ const rito = {
     es: {
         subtitle: 'web/app editorial project',
         overview: [
-            'Project created from scratch, in 4 days, in collaboration with Tadeau Carabias.',
-            'The challenge was to create a responsive online platform for a magazine to meet the needs and goals of users. The goal was to deliver a consistent, branded, and engaging experience and our main focus were on UI Design.'
+            'Proyecto creado desde cero, en 4 días, en colaboración con Tadeau Carabias.',
+            'El reto consistía en crear una plataforma online responsiva para una revista con el fin de satisfacer las necesidades y objetivos de los usuarios. El objetivo era ofrecer una experiencia consistente, de marca y atractiva y nuestro enfoque principal fue el diseño de la interfaz de usuario.'
         ],
-        scope: 'User Research, analyzing, user flow, information architecture, wireframing, mock-up, design system, prototyping, and usability testing.',
-        categories: 'Website/Mobile ',
+        scope: 'Investigación de usuarios, análisis, flujo de usuarios, arquitectura de la información, wireframing, mock-up, sistema de diseño, prototipado y pruebas de usabilidad.',
+        categories: 'Website/Movil',
         post: [
             {
                 sectionNumber: '01',
@@ -138,14 +138,14 @@ const rito = {
                     {
                         type: 'paragraph',
                         content:
-                            'Rito is an emergent art magazine to promote new and upcoming artists and review the work of established ones, with the intent to democratize art and enable easier access for everyone both artist and viewers',
+                            'Rito es una revista de arte emergente para promocionar a los nuevos artistas y revisar el trabajo de los establecidos, con la intención de democratizar el arte y permitir un acceso más fácil para todos, tanto para los artistas como para los espectadores.',
                         size: '200',
-                        weight: '400',
+                        weight: '500',
                         style: 'normal'
                     },
                     {
                         type: 'image',
-                        content: { src: rito1, alt: 'Rito 1' }
+                        content: { src: rito1, alt: 'Rito 1', size: 'auto' }
                     }
                 ]
             },
@@ -156,17 +156,17 @@ const rito = {
                     {
                         type: 'paragraph',
                         content:
-                            'At first, we searched the web for different digital and physical magazines to understand how this type of market works and what people want to consume when they decide to read an art magazine.',
+                            'Al principio, buscamos en la web diferentes revistas digitales y físicas para entender cómo funciona este tipo de mercado y qué es lo que la gente quiere consumir cuando decide leer una revista de arte.',
                         size: '200',
-                        weight: '400',
+                        weight: '500',
                         style: 'normal'
                     },
                     {
                         type: 'paragraph',
                         content:
-                            'After visiting and comparing different sites, the information they offer, how content is organized, their aesthetics, their accessibility, and even the frequency of content, we started thinking of questions for the interviews that would allow us to achieve our goal.',
+                            'Después de visitar y comparar diferentes sitios, la información que ofrecen, cómo se organizan los contenidos, su estética, su accesibilidad e incluso la frecuencia de los contenidos, empezamos a pensar en preguntas para las entrevistas que nos permitieran alcanzar nuestro objetivo.',
                         size: '200',
-                        weight: '400',
+                        weight: '500',
                         style: 'normal'
                     },
                     {
@@ -178,9 +178,9 @@ const rito = {
                     },
                     {
                         type: 'paragraph',
-                        content: 'We interviewed five people and this is what we learned from them.',
+                        content: 'Hemos entrevistado a cinco personas y esto es lo que hemos aprendido de ellas.',
                         size: '200',
-                        weight: '400',
+                        weight: '500',
                         style: 'normal'
                     },
                     {
@@ -190,15 +190,15 @@ const rito = {
                     {
                         type: 'paragraph',
                         content:
-                            'On the other hand, the surveys allow us to learn that magazines are a product that is not consumed daily, but more sporadically during the week. Also, 100% of the time, they are accessed by recommendation on social networks. Finally, people consume them because they look for inspiration and recommendations.',
+                            'Por otro lado, las encuestas nos permitieron conocer que las revistas son un producto que no se consumen a diario, sino más bien de forma esporádica durante la semana. Además, el 100% de las veces se accede a ellas por recomendación en las redes sociales. Por último, la gente las consume porque busca inspiración y recomendaciones.',
                         size: '200',
-                        weight: '400',
+                        weight: '500',
                         style: 'normal'
                     },
                     {
                         type: 'paragraph',
                         content:
-                            'In this stage of seeking to empathize with our users, we opened the door to Latin American art. We decided to create RITO to contribute to the visibility and democratization of emerging artists from Argentina and Brazil, countries that share territorial borders, proximity, and a long history of cooperation despite not sharing the same language.',
+                            'En esta etapa de búsqueda de empatía con nuestros usuarios, abrimos la puerta al arte latinoamericano. Decidimos crear RITO para contribuir a la visibilidad y democratización de los artistas emergentes de Argentina y Brasil, países que comparten fronteras territoriales, proximidad y una larga historia de cooperación a pesar de no compartir el mismo idioma.',
                         size: '400',
                         weight: '600',
                         style: 'normal'
@@ -212,9 +212,9 @@ const rito = {
                     {
                         type: 'paragraph',
                         content:
-                            'We choose a  colorful palette since colors are representative of Latin America.Our name was born after brainstorming a lot. We thought about our interviews, what we heard about them, and reflected on what we understood about art.',
+                            'Elegimos una paleta colorida ya que los colores son representativos de América Latina.Nuestro nombre nació después de una lluvia de ideas. Pensamos en nuestras entrevistas, en lo que oímos sobre ellas y reflexionamos sobre lo que entendíamos por arte.',
                         size: '200',
-                        weight: '400',
+                        weight: '500',
                         style: 'normal'
                     },
                     {
@@ -229,7 +229,7 @@ const rito = {
                 content: [
                     {
                         type: 'video',
-                        content: { src: rito6, alt: 'Rito 6' }
+                        content: { src: 'https://www.youtube.com/embed/PsXkEXV3cgA', alt: 'Rito 6' }
                     }
                 ]
             }